# swagger/swagger/swagger.yml
openapi: "3.0.0"
info:
  title: Users Service
  version: 1.0.0
  description: Users Service API documentation
servers:
  - url: http://localhost:5001
tags:
  - name: Auth
    description: Auth Operations
  - name: Users
    description: User Operations
  - name: Service Health
    description: Service Health Status
paths:
  /users-service/health:
    get:
      tags:
        - Service Health
      summary: Service health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /auth/login:
    post:
      tags:
        - Auth
      summay: Auth User
      description: Login User
      requestBody:
        description: Login data
        required: true
        content:
          application/json:
            schema:
              $ref: "#components/schemas/Login"
      responses:
        '200':
          description: User logued in.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    description: 'Auth Bearer Token'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#components/responses/BadRequest"
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#components/responses/NotFound"
  /auth/logout:
    get:
      tags:
        - Auth
      summay: Logout User
      description: Logout User
      security:
        - bearerAuth: []
      responses:
        '204':
          description: User logued out.

  /auth/status:
    get:
      tags:
        - Auth
      summay: Auth Status
      description: Get Auth Status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Auth Status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    description: 'Auth Status'
  /users:
    get:
      tags:
        - Users
      summary: List Users
      description: List Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List Users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#components/schemas/User"
    post:
      tags:
        - Users
      summary: Create User
      description: Creates a new User
      security:
        - bearerAuth: []
      requestBody:
        description: User data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User successfully created.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/User"
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#components/responses/BadRequest"
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get User
      description: Get an existing User
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: User id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User succesfullt retreived.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/User"
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#components/responses/NotFound"
    put:
      tags:
        - Users
      summary: Update User
      description: Update an existing User
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: User id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: User data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User successfully updated.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/User"
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#components/responses/BadRequest"
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#components/responses/NotFound"
    delete:
      tags:
        - Users
      summary: Delete User
      description: Delete an existing User
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: User id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: User successfully deleted.
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#components/responses/NotFound"
components:
  schemas:
    Login:
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: User email
        password:
          type: string
          description: User password
    UserCreate:
      allOf:
        - $ref: "#/components/requestBodies/UserBase"
        - required:
          - first_name
          - last_name
          - email
          - password
          properties:
            password:
              type: string
              description: User password
    UserUpdate:
      allOf:
        - $ref: "#/components/requestBodies/UserBase"
        - required:
          - first_name
          - last_name
          - email
    User:
      allOf:
        - $ref: "#/components/requestBodies/UserBase"
        - required:
          properties:
            id:
              type: integer
              format: int64
              description: User id
            first_name:
              type: string
              description: User first name
            last_name:
              type: string
              description: User last name
            email:
              type: string
              description: User email
            active:
              type: boolean
              description: User active status
    Health:
      properties:
        message:
          type: string
          description: Health status
  requestBodies:
    UserBase:
      properties:
        first_name:
          type: string
          description: User first name
        last_name:
          type: string
          description: User last name
        email:
          type: string
          description: User email
  responses:
    BadRequest:
      properties:
        message:
          type: string
          description: Error cause
    NotFound:
      properties:
        message:
          type: string
          description: Error cause

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer